APIGOV-18667 move new oas funcs to util
