+++ pkg/apic/specoas3processor.go
@@ -141,41 +141,3 @@ func (p *oas3SpecProcessor) parseURLsIntoEndpoints(defaultURL string, allURLs []
 
 	return endPoints, nil
 }
-
-// SetServers replaces the servers array on the Openapi3 object.
-func SetServers(hosts []string, spec *openapi3.T) {
-	var oas3Servers []*openapi3.Server
-	for _, s := range hosts {
-		oas3Servers = append(oas3Servers, &openapi3.Server{
-			URL: s,
-		})
-	}
-	if len(oas3Servers) > 0 {
-		spec.Servers = oas3Servers
-	}
-}
-
-// ParseOAS3 converts a JSON or YAML spec into an OpenAPI3 object
-func ParseOAS3(spec []byte) (*openapi3.T, error) {
-	oas3Obj, err := openapi3.NewLoader().LoadFromData(spec)
-	if err != nil {
-		return nil, err
-	}
-	if !strings.Contains(oas3Obj.OpenAPI, "3.") {
-		return nil, fmt.Errorf(oasParseError("3", ("'openapi' key is invalid.")))
-	}
-	if oas3Obj.Paths == nil {
-		return nil, fmt.Errorf(oasParseError("3", "'paths' key not found."))
-	}
-	if oas3Obj.Info == nil {
-		return nil, fmt.Errorf(oasParseError("3", "'info' key not found."))
-	}
-	if oas3Obj.Info.Title == "" {
-		return nil, fmt.Errorf(oasParseError("3", "'info.title' key not found."))
-	}
-	return oas3Obj, nil
-}
-
-func oasParseError(version string, msg string) string {
-	return fmt.Sprintf("invalid openapi %s specification. %s", version, msg)
-}
